requirements.txt
python-telegram-bot==20.3
Flask==2.1.2